refactor(daemon): fold plugin-asset-cache/plugin-preview-bakes into plugins/
Move the two top-level plugin-* leaf modules into the plugins/ folder so
plugin-owned modules live under one directory. Both files are pure leaf
modules (node/undici imports only), so the moves are byte-identical renames;
only the four import specifiers in server.ts and the two colocated tests are
repointed.

No public export surface change; no behavior change.
