+++ apps/daemon/src/plugins/plugin-asset-cache.ts
